Image file deletion

+++ app/Http/Controllers/SetImageController.php
@@ -22,7 +22,6 @@ class SetImageController extends Controller
         try {
             if ($request->hasFile('image')) {
                 $imageid = (string) Str::uuid();
-                error_log('image: ' . $imageid . ' set: ' . $setid);
                 Storage::disk('local')->put('images/sets/' . $imageid, $request->file('image'));
                 return SetImage::create([
                     'imageid' => $imageid,
@@ -32,7 +31,6 @@ class SetImageController extends Controller
                 return response()->json(['error' => 'Could not find attached file.'], 400);
             }
         } catch (Exception $e) {
-            error_log($e);
             return response()->json(['error' => 'Error uploading file.'], 500);
         }
     }
@@ -55,6 +53,10 @@ class SetImageController extends Controller
     public static function destroySetImage(Request $request)
     {
         $imageid = $request->imageid;
-        return SetImage::destroy($imageid);
+        if (!$imageid) return response()->json(['message' => 'Image not found'], 404);
+
+        Storage::disk('local')->delete('images/sets/' . $imageid);
+        SetImage::destroy($imageid);
+        return response()->json(['image' => $imageid], 200);
     }
 }
